Move command trait in to its own module, hide command boilerplate in the subscriber, add a get_subscribers method
