+++ src/events/mod.rs
@@ -2,5 +2,5 @@
 pub mod bus;
 pub mod commands;
 pub mod subscribers;
-pub mod command;
 pub mod broadcast;
+pub mod command;
