+++ src/events/subscribers/mod.rs
@@ -1,1 +1,2 @@
 ﻿pub mod chat;
+pub mod command;
